MemberController add exceptions

+++ src/main/java/com/alkemy/ong/controllers/MemberController.java
@@ -3,6 +3,7 @@ package com.alkemy.ong.controllers;
 import com.alkemy.ong.dtos.MemberDescriptionDTO;
 import com.alkemy.ong.dtos.MemberRequestDTO;
 import com.alkemy.ong.exceptions.DataAlreadyExistException;
+import com.alkemy.ong.exceptions.InvalidParameterException;
 import com.alkemy.ong.exceptions.NotFoundException;
 import com.alkemy.ong.services.MemberService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,25 +25,26 @@ public class MemberController {
     }
 
     @PostMapping("")
-    public ResponseEntity<?> save(@Valid @RequestBody MemberRequestDTO request){
+    public ResponseEntity<?> save(@Valid @RequestBody MemberRequestDTO request) throws DataAlreadyExistException{
 
-        try{
-            MemberDescriptionDTO response = memberService.create(request);
-            return ResponseEntity.ok(response);
-
-        }catch (DataAlreadyExistException ex){
-            return ResponseEntity.badRequest().build();
-        }
+        MemberDescriptionDTO response = memberService.create(request);
+        return ResponseEntity.ok(response);
 
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> delete(@PathVariable Long id){
-        try{
-            memberService.delete(id);
-        }catch(NotFoundException ex){
-            return new ResponseEntity(HttpStatus.BAD_REQUEST);
-        }
+    public ResponseEntity<?> delete(@PathVariable Long id) throws NotFoundException{
+        if(id == null || id.equals(0L)) throw new InvalidParameterException("Invalid id");
+
+        memberService.delete(id);
         return ResponseEntity.ok("Member successfuly deleted");
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<?> update(@PathVariable Long id, @Valid @RequestBody MemberRequestDTO request) throws NotFoundException{
+
+        memberService.update(request,id);
+
+        return ResponseEntity.ok().build();
     }
 }
